Bump ftw library to 0.0.46 to fix broken build and tests.
The 2.5.0 release of the addressable library made on 2016-11-04
started requiring Ruby 2.0 compatibility (via the introduction
of a dependency to the public_suffix library), so an attempt to
do a bundle installation for this plugin started to result in
the following:

    $ bundle install
    Fetching gem metadata from https://rubygems.org/.........
    Fetching version metadata from https://rubygems.org/..
    Fetching dependency metadata from https://rubygems.org/.
    Resolving dependencies.....................................
    public_suffix-2.0.4 requires ruby version >= 2.0, which is
    incompatible with the current version, ruby 1.9.3p551 (jruby
    1.7.26)

Since the ftw library that dragged in addressable hadn't locked
down its dependency it, too, effectively started requiring Ruby 2.0.
This is fixed in ftw 0.0.46 so by requiring at least that version we
make sure we don't get a too new version of the addressable library.

The new ftw also includes a bunch of other changes, mainly ones
related to TLS (like an updated CA certificate list), but hardly
anything that should have a detrimental affect on this plugin.